Update dependencies and modify schema in Phiture Braze Connector
This commit updates various dependencies and modifies the schema in the Phiture Braze Connector. It specifically changes the python base image, upgrades the airbyte-cdk and pytest versions, and replaces 'source-acceptance-test' with 'connector-acceptance-test'. The 'size' field type in the schema for segments_data_series has also been changed from 'integer' to 'number'.